requirements.txt
matplotlib>=3.2.2
torch>=1.7.0
Pillow>=7.1.2
numpy>=1.18.5
opencv-python>=4.1.1
requests>=2.23.0
scipy>=1.4.1
pandas>=1.1.4
seaborn>=0.11.0
tqdm>=4.64.0
torchvision>=0.8.1
PyYAML>=5.3.1
psutil>=5.9.4
streamlit>=1.19.0
streamlit_drawable_canvas>=0.9.2
ultralytics>=8.0.105
requests>=2.31.0
websockets>=12.0